fix(studio): Align on test suffix
