+++ web/packages/common/src/components/CodeEditor/util/editor.spec.ts
