+++ web/packages/common/src/components/SliderWithTextInput/SliderWithTextInput.spec.tsx
