+++ web/packages/common/src/components/TrainingParameterSlider/TrainingParameterSlider.spec.tsx
