+++ web/packages/common/src/namedEntity.spec.ts
