+++ web/packages/studio/src/components/dataViews/SafeSynthesizerJobsDataView/DeleteJobModal.spec.tsx
