+++ web/packages/studio/src/routes/SafeSynthesizerNewRoute/index.spec.tsx
